Refactor: Move backend code to designated classes.
