+++ app/src/main/java/Frontend/AppFrame.java
@@ -4,6 +4,8 @@
  */
 package Frontend;
 
+import Backend.InputListener;
+
 /**
  *
  * @author austinkwon
@@ -16,6 +18,7 @@ public class AppFrame extends javax.swing.JFrame {
     public AppFrame() {
         System.out.println("AppFrame() Constructor called!");
         initComponents();
+        buttonSubmit.addActionListener(new InputListener());
     }
 
     /**
@@ -23,7 +26,7 @@ public class AppFrame extends javax.swing.JFrame {
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
-    @SuppressWarnings("unchecked")
+
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
@@ -559,6 +562,7 @@ public class AppFrame extends javax.swing.JFrame {
     private void buttonSubmitActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonSubmitActionPerformed
         // TODO add your handling code here:
         System.out.println("Submit button pressed!");
+
     }//GEN-LAST:event_buttonSubmitActionPerformed
 
     /**
